+++ src/main/java/lermitage/intellij/extratci/BundledIcon.java
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: MIT
+
+package lermitage.intellij.extratci;
+
+/**
+ * An icon that comes with Extra Icons plugin.
+ */
+public record BundledIcon(String iconPath, String description) {
+}